// client/src/module/Admin/admin.validator.tsx
export interface listUserProps {
    _id: string,
    name: string,
    email: string,
    role: string,
    phone: string,
    isVerified: boolean,
    isBan: boolean,
    avatar: {
        public_id: string,
        secure_url: string,
        optimizedUrl: string
    },
    sellerProfile: {
        companyName: string,
        gstNumber: string,
        bio: string,
        address: string,
        rating: number,
        totalReviews: number
    },
    addresses: {
        label: string,
        fullName: string,
        phone: number,
        line1: string,
        line2: string,
        city: string,
        state: string,
        postalCode: string,
        country: string,
        isDefault: string
    }[],
    favourites: [string]
}

export interface BannerProps {
    _id: string;
    title: string;
    subtitle: string;
    type: string;
    startAt: string; // ISO date string
    endAt: string;   // ISO date string
    isActive: boolean;
    priority: number;
    createdAt: string;
    updatedAt: string;
    __v: number;
    image: {
        public_id: string;
        secure_url: string;
    };
}

export interface CouponProps {
    _id: string;
    code: string;
    discountType: 'percentage' | 'fixed'; // assuming you may have fixed amount discount as well
    discountValue: number;
    validFrom: string; // or Date if you plan to convert it
    validUntil: string; // or Date
    isActive: boolean;
    applicableCategories: string[];
    __v: number;
}
// ...
export interface BannerResponse {
    code: number;
    data: BannerProps[];
    message: string;
    options: {
        total: number;
    };
    status: string;
}

export interface CouponResponse {
    code: number;
    data: CouponProps[];
    message: string;
    options: {
        total: number;
    };
    status: string;
}

export interface UserResponse {
    code: number;
    data: listUserProps[];
    message: string;
    options: {
        total: number;
    };
    status: string;
}

export interface CategoryResponse {
    code: number;
    data: ListCategoryDetails[];
    message: string;
    options: {
        total: number;
    };
    status: string;
}